+++ Runtime/BaselibNetworkInterface.cs
@@ -759,10 +759,11 @@ namespace Unity.Networking.Transport
                 messagePtr,
                 1u,
                 &error);
-            if (error.code != ErrorCode.Success)
+            if (error.code != ErrorCode.Success || count != 1u)
             {
                 baselib->m_PayloadsTx.ReleaseHandle(index);
-                return (int)error.code == -1 ? -1 : -(int)error.code;
+                // This is the likeliest cause for this error.
+                return (int)Error.StatusCode.NetworkSendQueueFull;
             }
             return handle.size;
         }